pyproject.toml
[project]
name = "humanlayer"
version = "0.1.31"
description = "A general framework for simulating human user for agentic tasks."
readme = "README.md"
authors = [{ name = "Oliver Li", email = "aochongli@cs.cornell.edu" }]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.7",
    "shortuuid>=1.0.13",
    "typer>=0.16.0",
    "requests>=2.32.4",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "toml>=0.10.2",
    "tenacity>=9.1.2",
    "openai<1.100.0",
    "python-dotenv>=1.1.1",
    "litellm>=1.79.0",
    "jinja2>=3.1.6",
    "dirhash>=0.5.0",
    "dockerfile-parse>=2.0.1",
    "e2b>=2.4.2",
    "datasets>=4.4.1",
    "runloop-api-client>=1.2.0",
    "daytona>=0.121.0",
    "kubernetes>=32.0.0",
    "claude-agent-sdk>=0.1.17",
    "modal>=1.3.0.post1",
]

[project.scripts]
humanlayer = "humanlayer.cli.main:app"
hr = "humanlayer.cli.main:app"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dashboard = ["pandas>=2.3.3", "plotly>=6.4.0", "streamlit>=1.50.0"]
dev = [
    "ipykernel>=6.30.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
]

[tool.ruff.lint.isort]
known-first-party = ["humanlayer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--strict-markers", "--strict-config"]
markers = [
    "asyncio: marks tests as async tests",
    "runtime: marks tests as runtime tests (may need docker)",
    "unit: marks tests as unit tests (fast, no external dependencies)",
    "integration: marks tests as integration tests (requires external services, could be mocked)",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/humanlayer"]
omit = ["*/tests/*", "*/__pycache__/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]

[tool.ty.src]
include = ["src/humanlayer"]
